+++ src/components/form/form.js
@@ -1,14 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Button from './../button';
-import I18n from "i18n-js";
+import CancelButton from './cancel-button';
+import SaveButton from './save-button';
 import Serialize from "form-serialize";
 import classNames from 'classnames';
 import { validProps } from '../../utils/ether';
-import { assign } from 'lodash';
 import { tagComponent } from '../../utils/helpers/tags';
-
-import FormSummary from './form-summary';
 
 /**
  * A Form widget.
@@ -159,11 +156,11 @@ class Form extends React.Component {
     save: PropTypes.bool,
 
     /**
-     * Additional actions rendered next to the save and cancel buttons
-     *
-     * @property additionalActions
-     * @type {String|JSX}
-     */
+    * Additional actions rendered next to the save and cancel buttons
+    *
+    * @property additionalActions
+    * @type {String|JSX}
+    */
     additionalActions: PropTypes.node,
 
     /**
@@ -180,7 +177,15 @@ class Form extends React.Component {
      * @property activeInput
      * @type {Node}
      */
-    activeInput: PropTypes.node
+    activeInput: PropTypes.node,
+
+    /**
+     * Override Save Button
+     *
+     * @property customSaveButton
+     * @type {Node}
+     */
+    customSaveButton: PropTypes.node,
   }
 
   static defaultProps = {
@@ -189,7 +194,8 @@ class Form extends React.Component {
     cancel: true,
     save: true,
     saving: false,
-    validateOnMount: false
+    validateOnMount: false,
+    customSaveButton: null
   }
 
   static childContextTypes = {
@@ -482,6 +488,91 @@ class Form extends React.Component {
   }
 
   /**
+   * Gets the cancel button for the form
+   *
+   * @method cancelButton
+   * @return {Object} JSX cancel button
+   */
+  cancelButton = () => {
+    if (!this.props.cancel) { return null; }
+
+    let cancelProps = {
+      cancelText: this.props.cancelText,
+      cancelClick: this.cancelForm,
+      ...this.props.cancelButtonProps
+    };
+
+    return (
+      <CancelButton
+        data-element='cancel'
+        { ...cancelProps }
+      />
+    );
+  }
+
+  /**
+   * Gets any additional actions passed into the form
+   *
+   * @method additionalActions
+   * @return {Object} JSX
+   */
+  get additionalActions() {
+    if (!this.props.additionalActions) { return null; }
+
+    return (
+      <div className='carbon-form__additional-actions' >
+        { this.props.additionalActions }
+      </div>
+    );
+  }
+
+  /**
+   * The default Save button for the form
+   *
+   * @method defaultSaveButton
+   * @return {Object} JSX
+   */
+  defaultSaveButton = () => {
+    return(
+      <SaveButton
+        saveButtonProps={ this.props.saveButtonProps }
+        saveText={ this.props.saveText }
+        saving={ this.props.saving }
+        errors={ this.state.errorCount }
+        warnings={ this.state.warningCount }
+      />
+    );
+  }
+
+  /**
+   * Returns a custom save button if passed in
+   * the default if not
+   *
+   * @method saveButton
+   * @return {Object} JSX
+   */
+  saveButton = () => {
+    if (!this.props.save) { return null; }
+
+    return this.props.customSaveButton ? this.props.customSaveButton : this.defaultSaveButton();
+  }
+
+  /**
+   * Returns the buttons for the form
+   *
+   * @method buttons
+   * @return {Object} JSX
+   */
+  buttons = () => {
+    return (
+      <div>
+        { this.saveButton() }
+        { this.cancelButton() }
+      </div>
+    );
+  }
+
+  /**
    * Main class getter
    *
    * @method mainClasses
@@ -494,60 +585,16 @@ class Form extends React.Component {
     );
   }
 
+  /**
+   * Button class getter
+   *
+   * @method buttonClasses
+   * @return {String} Main className
+   */
   get buttonClasses() {
     return classNames(
       'carbon-form__buttons',
       `carbon-form__buttons--${ this.props.buttonAlign }`
-    );
-  }
-
-  /**
-   * Gets the cancel button for the form
-   *
-   * @method cancelButton
-   * @return {Object} JSX cancel button
-   */
-  get cancelButton() {
-    let cancelClasses = "carbon-form__cancel",
-        cancelProps = assign({}, this.props.cancelButtonProps, { type: 'button', onClick: this.cancelForm });
-
-    return (<div className={ cancelClasses }>
-      <Button { ...cancelProps } data-element='cancel'>
-        { this.props.cancelText || I18n.t('actions.cancel', { defaultValue: 'Cancel' }) }
-      </Button>
-    </div>);
-  }
-
-  get additionalActions() {
-    if (!this.props.additionalActions) { return null; }
-
-    return (
-      <div className='carbon-form__additional-actions' >
-        { this.props.additionalActions }
-      </div>
-    );
-  }
-
-  /**
-   * Gets the save button for the form
-   * @method saveButton
-   * @return {Object} JSX save button
-   */
-  get saveButton() {
-    let saveClasses = classNames(
-          "carbon-form__save", {
-            "carbon-form__save--invalid": this.state.errorCount || this.state.warningCount
-          }
-        ),
-        saveProps = assign({}, this.props.saveButtonProps, { as: 'primary', disabled: this.props.saving });
-
-    return (
-      <div className={ saveClasses }>
-        <FormSummary errors={ this.state.errorCount } warnings={ this.state.warningCount } />
-        <Button { ...saveProps } data-element='save'>
-          { this.props.saveText || I18n.t('actions.save', { defaultValue: 'Save' }) }
-        </Button>
-      </div>
     );
   }
 
@@ -558,16 +605,6 @@ class Form extends React.Component {
    * @return {Object} JSX form
    */
   render() {
-    let cancelButton, saveButton;
-
-    if (this.props.cancel) {
-      cancelButton = this.cancelButton;
-    }
-
-    if (this.props.save) {
-      saveButton = this.saveButton;
-    }
-
     return (
       <form onSubmit={ this.handleOnSubmit } { ...this.htmlProps() } ref="form" { ...tagComponent('form', this.props) }>
         { generateCSRFToken(this._document) }
@@ -575,8 +612,7 @@ class Form extends React.Component {
         { this.props.children }
 
         <div className={ this.buttonClasses }>
-          { saveButton }
-          { cancelButton }
+          { this.buttons() }
           { this.additionalActions }
         </div>
       </form>